requirements.txt
dill==0.3.6
multiprocess==0.70.14
nestd==0.3.1
pydantic==1.10.7
python-dotenv==1.0.0
robyn==0.29.0
typing_extensions==4.5.0
uvloop==0.17.0
watchdog==2.2.1
